Beta.

+++ lib/src/mouse/_apply_packet.c
@@ -7,7 +7,7 @@
  */
 #include "common.h"
 
-void _mouse_apply_packet(mouse_t *mouse, int8_t dx, int8_t dy, uint8_t buttons) {
+void _mouse_apply_packet(mouse_t *mouse, int16_t dx, int16_t dy, uint8_t buttons) {
     mouse->dx += dx;
     mouse->dy += dy;
     mouse->x += dx;
